Change a few concerns to use framework versions
Currently we have our own copy of

    InteractsWithDatabase
    InteractsWithConsole
    InteractsWithAuthentication
    InteractsWithSession
    InteractsWithExceptionHandling

That are the same as the core versions (but vary between 5.3, 5.4 and 5.5 (for example in 5.5 the functions for InteractsWithDatabase are different then the ones in this repo)). We should use the core versions for these rather then manually using our own.

Unless I am missing something?