Prevent installs of PHPUnit 8
It turns out that installs of PHPUnit 8 are currently not possible because we override the setUp method and its signature has changed. This commit reverst the ability to install PHPUnit 8 since it won't work anyway.

In a follow-up commit I'll send in support for PHPUnit 8 on the next release branch.